fix home screen ,add voice

+++ app/src/main/java/com/routesme/vehicles/data/model/BusPaymentProcessModel.kt
@@ -2,7 +2,7 @@ package com.routesme.vehicles.data.model
 
 import com.routesme.vehicles.BuildConfig
 
-data class BusPaymentProcessCredentials(var api_key: String = BuildConfig.NEW_API_KEY, var api_secret: String = BuildConfig.NEW_API_SECRET, var SecondID: String? = null, var PaymentCode: String? = null, var Value: Int? = null, var UserID: String? = null)
+data class BusPaymentProcessCredentials(var api_key: String = BuildConfig.NEW_API_KEY, var api_secret: String = BuildConfig.NEW_API_SECRET, var SecondID: String? = null, var PaymentCode: String? = null, var Value: Double? = null, var UserID: String? = null)
 
 data class BusPaymentProcessDTO(val status: Boolean = false, val description: Any? = null)
 data class BusPaymentProcessSuccessDTO (val message: String?  = null, val status: Boolean = false)
